requirements.txt
streamlit
numpy
pandas
scikit-learn
tweepy
plotly.express
pytz